+++ src/encoded/static/components/browse/browse-view/BrowseProtectedDonor.js
@@ -7,7 +7,6 @@ import { BrowseViewControllerWithSelections } from '../../static-pages/component
 import { BrowseViewAboveFacetListComponent } from './BrowseViewAboveFacetListComponent';
 import { BrowseViewAboveSearchTableControls } from './BrowseViewAboveSearchTableControls';
 import {
-    BROWSE_STATUS_FILTERS,
     BROWSE_LINKS,
     NoResultsBrowseModal,
 } from '../BrowseView';
@@ -18,6 +17,7 @@ import {
     customRenderDetailPane,
     createBaseDonorColumnExtensionMap,
 } from './BrowseDonorBase';
+import { buildDonorPeekMetadataHref } from './BrowseDonorPeekMetadata';
 
 export { formatTissueData, formatAssayData } from './BrowseDonorBase';
 
@@ -104,9 +104,7 @@ const BrowseProtectedDonorSearchTable = (props) => {
             ...context,
             clear_filters: BROWSE_LINKS.protected_donor,
         },
-        customColumnSearchHref: (result) =>
-            `/peek-metadata/?additional_facet=file_size&${BROWSE_STATUS_FILTERS}&dataset!=No+value&type=File&donors.display_title=` +
-            result?.display_title,
+        customColumnSearchHref: buildDonorPeekMetadataHref,
         defaultColAlignment: 'text-left',
     };
 
@@ -126,9 +124,7 @@ const BrowseProtectedDonorSearchTable = (props) => {
         createBrowseProtectedDonorColumnExtensionMap(selectedFileProps);
 
     const facetListSortFxns = {
-        hardy_scale: (a, b) => {
-            return a.key - b.key;
-        },
+        hardy_scale: (a, b) => a.key - b.key,
     };
 
     return (